#88 Add retries for rabbitmq listeners

+++ services/employee/src/main/java/com/senla/worklog/reminder/employee/adapter/out/jpa/RegionJpaAdapter.java
@@ -27,7 +27,9 @@ public class RegionJpaAdapter implements RegionJpaPort {
 
     @Override
     public void deleteRegionById(UUID id) {
-        regionRepository.deleteById(id);
+        if (regionRepository.existsById(id)) {
+            regionRepository.deleteById(id);
+        }
     }
 
     private Region saveRegionInternal(Region region) {
